+++ pages/write/index.page.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { TextEditor, CreateArticleModal } from 'src/components';
-import { TitleInput, Button } from 'src/components';
+import { TitleInput } from 'src/components';
 import styled from '@emotion/styled';
 
 const WritePage = () => {
@@ -20,14 +20,6 @@ const WritePage = () => {
           articleElements={articleElements}
         />
       )}
-      <ButtonWrapper>
-        <Button
-          onClick={handleOpenModal}
-          label="나가기"
-          buttonType="tertiary"
-        />
-        <Button onClick={handleOpenModal} label="완료" buttonType="secondary" />
-      </ButtonWrapper>
       <TitleWrapper>
         <TitleInput onChange={setTitle} />
       </TitleWrapper>
@@ -38,13 +30,6 @@ const WritePage = () => {
 };
 
 export default WritePage;
-
-const ButtonWrapper = styled.div`
-  padding: 25px;
-  display: flex;
-  justify-content: flex-end;
-  gap: 16px;
-`;
 
 const TitleWrapper = styled.div`
   width: 100%;
