+++ src/components/headerBar/HeaderBar.tsx
@@ -2,64 +2,18 @@ import React from 'react';
 import Link from 'next/link';
 import { useRouter } from 'next/router';
 import styled from '@emotion/styled';
-
-import { HeaderBarItem } from './HeaderBarItem';
+import { Menus } from './Menus';
+import { WritePageMenus } from './WritePageMenus';
 
 export const HeaderBar = () => {
   const router = useRouter();
-  const items = [
-    {
-      id: 'article',
-      name: 'Article',
-      title: '개발 아티클 리스트 페이지입니다.',
-      path: '/article',
-    },
-    {
-      id: 'book',
-      name: 'Book',
-      title: '독후감 아티클 리스트 페이지입니다.',
-      path: '/dev',
-    },
-    {
-      id: 'essay',
-      name: 'Essay',
-      title: '에세이 리스트 페이지입니다.',
-      path: '/dev',
-    },
-    {
-      id: 'quotes',
-      name: 'Quotes',
-      title: '인용구 리스트 페이지입니다.',
-      path: '/dev',
-    },
-    {
-      id: 'about',
-      name: 'About',
-      title: '소개페이지입니다.',
-      path: '/dev',
-    },
-  ];
 
   return (
     <Container>
       <Link href="/">
         <Title>recoen.</Title>
       </Link>
-      <nav>
-        <Wrapper>
-          {items.map((item) => {
-            return (
-              <HeaderBarItem
-                key={item.id}
-                title={item.title}
-                name={item.name}
-                path={item.path}
-                isActive={router.pathname.includes(item.path)}
-              />
-            );
-          })}
-        </Wrapper>
-      </nav>
+      {router.pathname == '/write' ? <WritePageMenus /> : <Menus />}
     </Container>
   );
 };
@@ -83,10 +37,4 @@ const Title = styled.a`
   font-weight: 600;
   color: #ffffff;
   cursor: pointer;
-`;
-
-const Wrapper = styled.ul`
-  display: flex;
-  gap: 40px;
-  list-style: none;
 `;
